v2ray-vpn/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation and serialization
pydantic==2.5.0

# V2Ray integration
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1

# HTTP client and utilities
httpx==0.25.2
aiofiles==23.2.1

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring and metrics
prometheus-client==0.19.0

# Configuration and environment
python-dotenv==1.0.0

# Utilities
click==8.1.7
jinja2==3.1.2 